LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controller-on-reload.https-expected.txt


PASS controller is set upon reload after registration
